app/requirements.txt
asgiref==3.7.2
Django==4.2.3
django-debug-toolbar==4.1.0
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
Markdown==3.4.3
psycopg2==2.9.6
pyaml==23.7.0
Pygments==2.15.1
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
django-rest-swagger
